small pybind conflict fixes in unittests
