+++ src/gsMaterialMatrixBase_.cpp
@@ -39,22 +39,6 @@ namespace gismo
     ;
   }
 
-  void pybind11_init_gsMaterialMatrixBaseDim2(py::module &m)
-  {
-    using Base  = gsMaterialMatrixBase<real_t>;
-    using Class = gsMaterialMatrixBaseDim<2,real_t>;
-    py::class_<Class,Base>(m, "gsMaterialMatrixBaseDim2")
-    ;
-  }
-
-  void pybind11_init_gsMaterialMatrixBaseDim3(py::module &m)
-  {
-    using Base  = gsMaterialMatrixBase<real_t>;
-    using Class = gsMaterialMatrixBaseDim<3,real_t>;
-    py::class_<Class,Base>(m, "gsMaterialMatrixBaseDim3")
-    ;
-  }
-
   #endif
 
 }
